+++ src/printer.rs
@@ -7,8 +7,6 @@ use nu_ansi_term::Style;
 
 use bytesize::ByteSize;
 
-use console::AnsiCodeIterator;
-
 use syntect::easy::HighlightLines;
 use syntect::highlighting::Color;
 use syntect::highlighting::Theme;
@@ -33,8 +31,22 @@ use crate::line_range::RangeCheckResult;
 use crate::preprocessor::{expand_tabs, replace_nonprintable};
 use crate::style::StyleComponent;
 use crate::terminal::{as_terminal_escaped, to_ansi_color};
-use crate::vscreen::AnsiStyle;
+use crate::vscreen::{AnsiStyle, EscapeSequence, EscapeSequenceIterator};
 use crate::wrapping::WrappingMode;
+
+const ANSI_UNDERLINE_ENABLE: EscapeSequence = EscapeSequence::CSI {
+    raw_sequence: "\x1B[4m",
+    parameters: "4",
+    intermediates: "",
+    final_byte: "m",
+};
+
+const ANSI_UNDERLINE_DISABLE: EscapeSequence = EscapeSequence::CSI {
+    raw_sequence: "\x1B[24m",
+    parameters: "24",
+    intermediates: "",
+    final_byte: "m",
+};
 
 pub enum OutputHandle<'a> {
     IoWrite(&'a mut dyn io::Write),
@@ -287,6 +299,14 @@ impl<'a> InteractivePrinter<'a> {
         }
     }
 
+    fn get_header_component_indent_length(&self) -> usize {
+        if self.config.style_components.grid() && self.panel_width > 0 {
+            self.panel_width + 2
+        } else {
+            self.panel_width
+        }
+    }
+
     fn print_header_component_indent(&mut self, handle: &mut OutputHandle) -> Result<()> {
         if self.config.style_components.grid() {
             write!(
@@ -300,6 +320,30 @@ impl<'a> InteractivePrinter<'a> {
         } else {
             write!(handle, "{}", " ".repeat(self.panel_width))
         }
+    }
+
+    fn print_header_component_with_indent(
+        &mut self,
+        handle: &mut OutputHandle,
+        content: &str,
+    ) -> Result<()> {
+        self.print_header_component_indent(handle)?;
+        writeln!(handle, "{}", content)
+    }
+
+    fn print_header_multiline_component(
+        &mut self,
+        handle: &mut OutputHandle,
+        content: &str,
+    ) -> Result<()> {
+        let mut content = content;
+        let content_width = self.config.term_width - self.get_header_component_indent_length();
+        while content.len() > content_width {
+            let (content_line, remaining) = content.split_at(content_width);
+            self.print_header_component_with_indent(handle, content_line)?;
+            content = remaining;
+        }
+        self.print_header_component_with_indent(handle, content)
     }
 
     fn preprocess(&self, text: &str, cursor: &mut usize) -> String {
@@ -377,31 +421,32 @@ impl<'a> Printer for InteractivePrinter<'a> {
             }
         }
 
-        header_components.iter().try_for_each(|component| {
-            self.print_header_component_indent(handle)?;
-
-            match component {
-                StyleComponent::HeaderFilename => writeln!(
-                    handle,
-                    "{}{}{}",
-                    description
-                        .kind()
-                        .map(|kind| format!("{}: ", kind))
-                        .unwrap_or_else(|| "".into()),
-                    self.colors.header_value.paint(description.title()),
-                    mode
-                ),
-
+        header_components
+            .iter()
+            .try_for_each(|component| match component {
+                StyleComponent::HeaderFilename => {
+                    let header_filename = format!(
+                        "{}{}{}",
+                        description
+                            .kind()
+                            .map(|kind| format!("{}: ", kind))
+                            .unwrap_or_else(|| "".into()),
+                        self.colors.header_value.paint(description.title()),
+                        mode
+                    );
+                    self.print_header_multiline_component(handle, &header_filename)
+                }
                 StyleComponent::HeaderFilesize => {
                     let bsize = metadata
                         .size
                         .map(|s| format!("{}", ByteSize(s)))
                         .unwrap_or_else(|| "-".into());
-                    writeln!(handle, "Size: {}", self.colors.header_value.paint(bsize))
+                    let header_filesize =
+                        format!("Size: {}", self.colors.header_value.paint(bsize));
+                    self.print_header_multiline_component(handle, &header_filesize)
                 }
                 _ => Ok(()),
-            }
-        })?;
+            })?;
 
         if self.config.style_components.grid() {
             if self.content_type.map_or(false, |c| c.is_text()) || self.config.show_nonprintable {
@@ -521,7 +566,7 @@ impl<'a> Printer for InteractivePrinter<'a> {
             self.config.highlighted_lines.0.check(line_number) == RangeCheckResult::InRange;
 
         if highlight_this_line && self.config.theme == "ansi" {
-            self.ansi_style.update("^[4m");
+            self.ansi_style.update(ANSI_UNDERLINE_ENABLE);
         }
 
         let background_color = self
@@ -548,23 +593,17 @@ impl<'a> Printer for InteractivePrinter<'a> {
             let italics = self.config.use_italic_text;
 
             for &(style, region) in &regions {
-                let ansi_iterator = AnsiCodeIterator::new(region);
+                let ansi_iterator = EscapeSequenceIterator::new(region);
                 for chunk in ansi_iterator {
                     match chunk {
-                        // ANSI escape passthrough.
-                        (ansi, true) => {
-                            self.ansi_style.update(ansi);
-                            write!(handle, "{}", ansi)?;
-                        }
-
                         // Regular text.
-                        (text, false) => {
-                            let text = &*self.preprocess(text, &mut cursor_total);
+                        EscapeSequence::Text(text) => {
+                            let text = self.preprocess(text, &mut cursor_total);
                             let text_trimmed = text.trim_end_matches(|c| c == '\r' || c == '\n');
 
                             write!(
                                 handle,
-                                "{}",
+                                "{}{}",
                                 as_terminal_escaped(
                                     style,
                                     &format!("{}{}", self.ansi_style, text_trimmed),
@@ -572,9 +611,11 @@ impl<'a> Printer for InteractivePrinter<'a> {
                                     colored_output,
                                     italics,
                                     background_color
-                                )
+                                ),
+                                self.ansi_style.to_reset_sequence(),
                             )?;
 
+                            // Pad the rest of the line.
                             if text.len() != text_trimmed.len() {
                                 if let Some(background_color) = background_color {
                                     let ansi_style = Style {
@@ -592,6 +633,12 @@ impl<'a> Printer for InteractivePrinter<'a> {
                                 write!(handle, "{}", &text[text_trimmed.len()..])?;
                             }
                         }
+
+                        // ANSI escape passthrough.
+                        _ => {
+                            write!(handle, "{}", chunk.raw())?;
+                            self.ansi_style.update(chunk);
+                        }
                     }
                 }
             }
@@ -601,17 +648,11 @@ impl<'a> Printer for InteractivePrinter<'a> {
             }
         } else {
             for &(style, region) in &regions {
-                let ansi_iterator = AnsiCodeIterator::new(region);
+                let ansi_iterator = EscapeSequenceIterator::new(region);
                 for chunk in ansi_iterator {
                     match chunk {
-                        // ANSI escape passthrough.
-                        (ansi, true) => {
-                            self.ansi_style.update(ansi);
-                            write!(handle, "{}", ansi)?;
-                        }
-
                         // Regular text.
-                        (text, false) => {
+                        EscapeSequence::Text(text) => {
                             let text = self.preprocess(
                                 text.trim_end_matches(|c| c == '\r' || c == '\n'),
                                 &mut cursor_total,
@@ -654,7 +695,7 @@ impl<'a> Printer for InteractivePrinter<'a> {
                                     // It wraps.
                                     write!(
                                         handle,
-                                        "{}\n{}",
+                                        "{}{}\n{}",
                                         as_terminal_escaped(
                                             style,
                                             &format!("{}{}", self.ansi_style, line_buf),
@@ -663,6 +704,7 @@ impl<'a> Printer for InteractivePrinter<'a> {
                                             self.config.use_italic_text,
                                             background_color
                                         ),
+                                        self.ansi_style.to_reset_sequence(),
                                         panel_wrap.clone().unwrap()
                                     )?;
 
@@ -691,6 +733,12 @@ impl<'a> Printer for InteractivePrinter<'a> {
                                 )
                             )?;
                         }
+
+                        // ANSI escape passthrough.
+                        _ => {
+                            write!(handle, "{}", chunk.raw())?;
+                            self.ansi_style.update(chunk);
+                        }
                     }
                 }
             }
@@ -711,8 +759,8 @@ impl<'a> Printer for InteractivePrinter<'a> {
         }
 
         if highlight_this_line && self.config.theme == "ansi" {
-            self.ansi_style.update("^[24m");
-            write!(handle, "\x1B[24m")?;
+            write!(handle, "{}", ANSI_UNDERLINE_DISABLE.raw())?;
+            self.ansi_style.update(ANSI_UNDERLINE_DISABLE);
         }
 
         Ok(())